Refactor fusion recipe components  to prepare for fusion component finder
