+++ src/FusionRecipe/FusionRecipe.js
@@ -1,0 +1,7 @@
+import { FusionRecipeSkillFinder } from "./FusionRecipeSkillFinder.js";
+
+export function FusionRecipe() {
+  return (
+    <FusionRecipeSkillFinder />
+  )
+}
